Moved common classes for all streams into org.apache.pinot.plugin.stream
The classes SimpleAvroMessageDecoder and AvroRecordToPinotRowGenerator are
common to all streams using avro. For now, left these in kafka packages,until
we find a better home. Changed the class name to be org.apache.pinot.plugin.stream
